refactor(providers): isolate shared text model client
