+++ server/test/memory-extractor.test.mjs
@@ -6,10 +6,7 @@ import test from 'node:test'
 import { ConversationSync } from '../src/conversation/conversation-sync.mjs'
 import { FrontendMemoryService } from '../src/conversation/frontend-memory-service.mjs'
 import { MarkdownContextStore } from '../src/conversation/markdown-context-store.mjs'
-import {
-  MemoryExtractor,
-  createExtractorLlmCall,
-} from '../src/conversation/memory-extractor.mjs'
+import { MemoryExtractor } from '../src/conversation/memory-extractor.mjs'
 
 const OWNER = 'owner'
 const SESSION = 'main'
@@ -318,46 +315,4 @@ test('survives provider and malformed-output failures without throwing', async (
     await instance.maybeRun({ ownerId: OWNER, sessionId: SESSION })
     assert.equal(events.at(-1).op, 'error')
   }
-})
-
-test('createExtractorLlmCall posts to chat completions and surfaces errors', async () => {
-  assert.equal(createExtractorLlmCall({ baseUrl: 'https://example.com/v1', apiKey: '' }), null)
-  const requests = []
-  const llmCall = createExtractorLlmCall({
-    baseUrl: 'https://example.com/v1',
-    apiKey: 'test-key',
-    model: 'qwen-flash',
-    fetchImpl: async (url, options) => {
-      requests.push({ url, options })
-      return { ok: true, json: async () => ({ choices: [{ message: { content: '{}' } }] }) }
-    },
-  })
-  assert.equal(await llmCall({ system: 's', user: 'u' }), '{}')
-  assert.equal(requests[0].url, 'https://example.com/v1/chat/completions')
-  assert.equal(requests[0].options.headers.Authorization, 'Bearer test-key')
-  assert.equal(JSON.parse(requests[0].options.body).temperature, 0)
-
-  const retriedRequests = []
-  const compatible = createExtractorLlmCall({
-    baseUrl: 'https://example.com/v1',
-    apiKey: 'test-key',
-    model: 'strict-compatible-model',
-    fetchImpl: async (url, options) => {
-      retriedRequests.push({ url, options })
-      return retriedRequests.length === 1
-        ? { ok: false, status: 400, text: async () => 'unsupported temperature' }
-        : { ok: true, json: async () => ({ choices: [{ message: { content: '[]' } }] }) }
-    },
-  })
-  assert.equal(await compatible({ system: 's', user: 'u' }), '[]')
-  assert.equal(retriedRequests.length, 2)
-  assert.equal('temperature' in JSON.parse(retriedRequests[1].options.body), false)
-
-  const failing = createExtractorLlmCall({
-    baseUrl: 'https://example.com/v1',
-    apiKey: 'test-key',
-    model: 'qwen-flash',
-    fetchImpl: async () => ({ ok: false, status: 429, text: async () => 'rate limited' }),
-  })
-  await assert.rejects(() => failing({ system: 's', user: 'u' }), /request failed: 429 rate limited/)
 })
